Revert "Merge Model max length into main and rename evaluator (#98)"
This reverts commit 8ad27270249ae720077fd3b1fd6be09dceec5d3d.
