+++ prompt2model/evaluator/__init__.py
@@ -1,0 +1,6 @@
+"""Import evaluator classes."""
+from prompt2model.evaluator.base import Evaluator
+from prompt2model.evaluator.generate import Seq2SeqEvaluator
+from prompt2model.evaluator.mock import MockEvaluator
+
+__all__ = ("MockEvaluator", "Evaluator", "Seq2SeqEvaluator")
