--- prompt2model/model_evaluator/__init__.py
@@ -1,6 +1,0 @@
-"""Import evaluator classes."""
-from prompt2model.model_evaluator.base import Evaluator
-from prompt2model.model_evaluator.generate import Seq2SeqEvaluator
-from prompt2model.model_evaluator.mock import MockEvaluator
-
-__all__ = ("MockEvaluator", "Evaluator", "Seq2SeqEvaluator")
